Updates to CC transmon and support for leakage cal

+++ pycqed/measurement/Pulse_Generator.py
@@ -61,8 +61,8 @@ def block_pulse(amp, length, sampling_rate=2e8, delay=0, phase=0):
         phase in degrees
     '''
     nr_samples = (length+delay)*sampling_rate
-    delay_samples = delay*sampling_rate
-    pulse_samples = nr_samples - delay_samples
+    delay_samples = int(delay*sampling_rate)
+    pulse_samples = int(nr_samples - delay_samples)
     amp_I = amp*np.cos(phase*2*np.pi/360)
     amp_Q = amp*np.sin(phase*2*np.pi/360)
     block_I = amp_I * np.ones(pulse_samples)
